+++ changelog.md
@@ -6,5 +6,6 @@
 - Fixed profit calculations.
 - Switched GitHub Pages to Firebase.
 - Reduced power of "Samson 45K-FGXL".
+- Added Firebase analytics.
 
 ## v0.9 - MVP Pre Release - 30/01/2022